requirements.txt
alpaca-py==0.33.1
anthropic==0.52.0
apscheduler==3.10.4
pyyaml==6.0.2
feedparser==6.0.11
requests==2.32.3
pytz==2024.2
python-dateutil==2.9.0
yfinance>=0.2.40
